requirements.txt
pytesseract~=0.3.9
Pillow~=9.1.0